// renderer/src/app/shared/domain/metadata/search.enum.ts
export enum SubjectType {
  PAI = 'PAI',
  PAE = 'PAE',
  AS = 'AS',
  PG = 'PG',
  PF = 'PF',
  RUP = 'RUP',
  SW = 'SW',
  ALTRO = '',
}